pyproject.toml
[tool.poetry]
name = "payments_microservice"
version = "0.1.0"
description = ""
authors = ["AlexandreCotorobai <alexandrecotorobai@gmail.com>"]
readme = "README.md"

packages = [
    { include = "crud" },
    { include = "auth" },
    { include = "db" },
    { include = "models" },
    { include = "routers" },
    { include = "tests" }
]

[tool.poetry.dependencies]
python = "^3.10.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.1"
sqlalchemy = "^2.0.35"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
boto3 = "^1.35.38"
pymysql = "^1.1.1"
requests = "^2.32.3"
cryptography = "^43.0.1"
python-jose = "^3.3.0"
httpx = "^0.27.2"
tox = "^4.21.2"
testcontainers = "^4.8.1"
aio-pika = "^9.4.3"
stripe = "^11.1.1"

[tool.poetry.group.dev.dependencies]
coverage = "^7.6.2"
pytest-cov = "^5.0.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"